requirements.txt
ccTalkPython

pyserial >= 3.0